Move test folder into src This needs to be done because the Webpack settings don't allow anything outside src to be imported
