+++ packages/jbrowse-web/src/test/root.snap.1.json
