+++ packages/jbrowse-web/src/ui/App.test.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import App from './App'
 import JBrowse from '../JBrowse'
 import RootModelFactory from '../RootModelFactory'
-import snap1 from '../../test/root.snap.1.json'
+import snap1 from '../test/root.snap.1.json'
 import LinearGenomeView from '../plugins/LinearGenomeView/LinearGenomeView'
 import LinearGenomeViewModel from '../plugins/LinearGenomeView/model'
 import AppBar from './AppBar/AppBar'
